Move RecurringActionServiceTests to separate file
