+++ wire-ios-sync-engine/Tests/Source/UserSession/ZMUserSessionTests+RecurringActions.swift
@@ -61,60 +61,3 @@ class ZMUserSessionTests_RecurringActions: ZMUserSessionTestsBase {
         XCTAssertEqual(mockRecurringActionService.actions.count, 1)
     }
 }
-
-class RecurringActionServiceTests: ZMTBaseTest {
-
-    var sut: RecurringActionService!
-    let actionID = "11"
-    var actionPerformed = false
-    let mockUserDefaultsName = "mock"
-
-    override func setUp() {
-        super.setUp()
-
-        sut = RecurringActionService()
-        sut.userDefaultsName = mockUserDefaultsName
-    }
-
-    override func tearDown() {
-        sut = nil
-        actionPerformed = false
-
-        super.tearDown()
-    }
-
-    func testThatItPerformsAction() {
-        // given
-        sut.persistLastActionDate(for:actionID)
-        let action = RecurringAction(id: actionID, interval: 1, perform: { self.actionPerformed = true })
-        sut.registerAction(action)
-
-        // when
-        Thread.sleep(forTimeInterval: 2)
-        sut.performActionsIfNeeded()
-
-        // then
-        XCTAssertTrue(actionPerformed)
-
-        // tearDown
-        UserDefaults.standard.removePersistentDomain(forName: mockUserDefaultsName + actionID)
-    }
-
-    func testThatItDoesNotPerformAction_TimeHasNotExpired() {
-        // given
-        let action = RecurringAction(id: actionID, interval: 4, perform: { self.actionPerformed = true })
-        sut.registerAction(action)
-        sut.persistLastActionDate(for: actionID)
-
-        // when
-        Thread.sleep(forTimeInterval: 2)
-        sut.performActionsIfNeeded()
-
-        // then
-        XCTAssertFalse(actionPerformed)
-
-        // tearDown
-        UserDefaults.standard.removePersistentDomain(forName: mockUserDefaultsName + actionID)
-    }
-
-}
